move existing feature gates into feature package
